requirements.txt
# Core dependencies
sympy>=1.12
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pyarrow>=11.0.0  # For parquet support
pyyaml>=6.0
dill>=0.3.6  # Better serialization

# Progress and utilities
tqdm>=4.65.0

# Optional but recommended
ipython>=8.0.0
jupyter>=1.0.0
pytest>=7.0.0  # For testing
black>=23.0.0  # Code formatting
mypy>=1.0.0  # Type checking

# Performance monitoring (optional)
psutil>=5.9.0
memory_profiler>=0.60.0

# Add to requirements.txt

# Web interface
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# API interface  
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# CLI interface
click>=8.1.0
tabulate>=0.9.0
colorama>=0.4.6

# UI components
plotly>=5.0.0
dash>=2.0.0  # Optional for advanced dashboards